Refactor common code of pysdk in remote_thrift/utils.py
