+++ python/infinity_sdk/infinity/remote_thrift/table.py
@@ -26,8 +26,14 @@ from infinity.errors import ErrorCode
 from infinity.index import IndexInfo
 from infinity.remote_thrift.query_builder import Query, InfinityThriftQueryBuilder, ExplainQuery
 from infinity.remote_thrift.types import build_result
-from infinity.remote_thrift.utils import traverse_conditions, name_validity_check, select_res_to_polars
-from infinity.remote_thrift.utils import get_remote_constant_expr_from_python_value
+from infinity.remote_thrift.utils import (
+    traverse_conditions,
+    name_validity_check,
+    select_res_to_polars,
+    check_valid_name,
+    get_remote_constant_expr_from_python_value,
+    get_ordinary_info,
+)
 from infinity.table import ExplainType
 from infinity.common import ConflictType, DEFAULT_MATCH_VECTOR_TOPN
 from infinity.utils import deprecated_api
